+++ phpstan.neon
@@ -33,6 +33,8 @@ parameters:
          - '#Method Migrify\\SymfonyRouteUsage\\EntityRepository\\RouteVisitRepository\:\:findByRouteHash\(\) should return Migrify\\SymfonyRouteUsage\\Entity\\RouteVisit\|null but returns object\|null#'
          - '#Parameter \#1 \$error_handler of function set_error_handler expects \(callable\(int, string, string, int, array\)\: bool\)\|null, Closure\(int, string\)\: void given#'
          - '#Cannot cast array<string\>\|bool\|string\|null to string#'
-         - '#Class with base "FileFinder" name is already used in "PHPStan\\File\\FileFinder", "Migrify\\ClassPresence\\Finder\\FileFinder"\. Use unique name to make classes easy to recognize#'
          - '#Cannot cast array<string\>\|bool\|string\|null to int#'
          - '#Property PhpParser\\Node\\Stmt\\Namespace_\:\:\$stmts \(array<PhpParser\\Node\\Stmt\>\) does not accept array<int\|string, PhpParser\\Node\>#'
+         - '#Cognitive complexity for "Migrify\\ConfigTransformer\\FormatSwitcher\\PhpParser\\NodeFactory\\FluentClosureNamespaceNodeFactory\:\:(convertServiceOptionsToNodes|addServicesNodes)\(\)" is \d+, keep it under \d+#'
+         - '#Class cognitive complexity for "FluentClosureNamespaceNodeFactory" is \d+, keep it under 50#'
+         - '#Variable \$methodCall might not be defined#'
