+++ rector-ci.yaml
@@ -24,7 +24,8 @@ parameters:
         # unknown purpose
         - 'packages/diff-data-miner/src/Extractor/DefaultValueChangesExtractor.php'
         # 3rd party files
-        - 'packages/config-transformer/packages/format-switcher/src/Converter/YamlToPhpConverter.php'
+#        - 'packages/config-transformer/packages/format-switcher/src/Converter/YamlToPhpConverter.php'
+        - 'packages/config-transformer/packages/format-switcher/src/PhpParser/NodeFactory/FluentClosureNamespaceNodeFactory.php'
 
     exclude_rectors:
         # false positive on packages/symfony-route-usage/src/Database/TableInitiator.php
